Implement end-to-end example (KDSE-ACT-002)
- Add examples/complete/main.go demonstrating full architecture
- Add examples/complete/main_test.go with 3 passing tests
- Add docs/examples/end-to-end.md documentation
- Update roadmap to mark Milestone 1.1 and 1.2 as complete
- Update docs/README.md with Examples section
- Fix type errors in memory/memory.go (Float32 bits handling)
- Fix type error in models/models.go (elevation calculation)
- Fix type error in scheduler/scheduler.go (Model type)
- Add go.sum for dependencies

This implements KDSE-ACT-002: End-to-End Example as approved by operator.

+++ models/models.go
@@ -210,7 +210,7 @@ func (s *SunModel) Tick() {
 	}
 
 	// Calculate elevation from hour angle (simplified)
-	s.elevation = 45.0 * math.Sin(float64(s.solarHourAngle)*math.Pi/180.0)
+	s.elevation = float32(45.0 * math.Sin(float64(s.solarHourAngle)*math.Pi/180.0))
 	if s.elevation < 0 {
 		s.elevation = 0
 		s.irradiance = 0
